Seperated PopUp to two classes (S.PopUp, A,PopUp)
